Add payments seeds and while doing so make some ajustments (data type, extra columns in tasks, etc)

+++ server/db/schema/11_payments.sql
@@ -4,11 +4,11 @@ CREATE TABLE payments (
   user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
   tasker_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
   -- all values are expected in cents
-  subtotal INTEGER NOT NULL,
+  subtotal NUMERIC NOT NULL,
   discount INTEGER,
-  comission INTEGER NOT NULL,
-  tax INTEGER NOT NULL,
-  total INTEGER NOT NULL,
+  comission NUMERIC NOT NULL,
+  tax NUMERIC NOT NULL,
+  total NUMERIC NOT NULL,
   confirmed_at TIMESTAMP,
   confirmation_code VARCHAR(255),
   PRIMARY KEY (task_id, tasker_id, user_id)
